requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.5.1
shortuuid==1.0.13
sqlparse==0.4.4
tzdata==2023.3
wheel==0.43.0
